outfit-recommend-server/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
pydantic-settings==2.9.1
python-multipart==0.0.20
Pillow==11.2.1
torch==2.7.0
transformers==4.51.3
httpx==0.28.1
